Fix docerization

+++ repository/entry.go
@@ -27,10 +27,10 @@ values (:title, :author, :summary, :link, :published_at, :feed_id, :read_at, :cr
 
 type entryRepository struct {
 	db      *sqlx.DB
-	perPage int64
+	perPage uint64
 }
 
-func NewEntryRepository(db *sqlx.DB, perPage int64) *entryRepository {
+func NewEntryRepository(db *sqlx.DB, perPage uint64) *entryRepository {
 	return &entryRepository{
 		db:      db,
 		perPage: perPage,
@@ -55,7 +55,7 @@ func (r *entryRepository) GetByURL(ctx context.Context, url string) (Entry, erro
 
 func (r *entryRepository) ListForFeed(ctx context.Context, feedID, page int64) ([]Entry, error) {
 	entries := []Entry{}
-	if err := r.db.SelectContext(ctx, &entries, selectEntriesForFeedQuery, feedID, r.perPage, r.perPage*(page-1)); err != nil {
+	if err := r.db.SelectContext(ctx, &entries, selectEntriesForFeedQuery, feedID, r.perPage, r.perPage*uint64(page-1)); err != nil {
 		return nil, fmt.Errorf("cannot select entries: %w", err)
 	}
 	return entries, nil
